==> Code_R/2_Inspecting doc_id:city:state(+ZZ).R <==
# Inspecting doc_id, city, state(+ZZ)
# Crucial key to connect all tables !

# is doc_id similar to npi ? npi must have 10, while doc_id has 2+7. # is doc_id generated by data provider ?
which(dt_names %like% '.*doc_id*')
str_count(presc_detail$doc_id[1])
doc_id_string_counter <- function(x) {
                          doc_id_length <- lapply(x$doc_id, str_count)
                          less_than <- 0
                          greater_than <- 0
                          actual <- 0
                          for (i in 1:length(x$doc_id)) {
                            if (doc_id_length[[i]] < 9) {
                              less_than <- less_than + 1
                            } else if (doc_id_length[[i]] > 9) {
                              greater_than <- greater_than + 1
                            } else if (doc_id_length[[i]] == 9) {
                              actual <- actual + 1
                            }
                          }
                          if (actual == length(x$doc_id)) {
                            paste('Charater length is equal 9 for all doc_id')
                          } else {
                            c(less_than, greater_than)
                          }
                        }

doc_id_string_counter(presc_detail)
doc_id_string_counter(presc_sumry)
doc_id_string_counter(PUF_conso)
doc_id_string_counter(PUF_detail)
doc_id_string_counter(PUF_sumry)

lapply(dt_names[[1]], find_desc, match = 1)

# state & city in PUF_conso same as provider_* in others ? # need to look into it !
# create look-up table for states at a later stage !
all_names[all_names %like% '.*provider*']
which(dt_names %like% '.*provider*')
head(names(PUF_conso))

# Prescriber_Detailed is the main file because further profiling will be based on doc_ids conditionally pulled from this file
# how to handle state = 'ZZ' ? # foreign vs domestic ?
sum(presc_detail$nppes_provider_state == 'ZZ')
state_zz <- presc_detail[presc_detail$nppes_provider_state == 'ZZ']

unique(state_zz$doc_id)

select(presc_detail, doc_id, nppes_provider_state) %>% filter(doc_id == 'SR7876440') # No other state
select(presc_detail, doc_id, nppes_provider_state) %>% filter(doc_id == 'ZH4066278') # No other state
select(presc_detail, doc_id, nppes_provider_state) %>% filter(doc_id == 'PI6937446') # No other state
select(presc_detail, doc_id, nppes_provider_state) %>% filter(doc_id == 'FP6297812') # No other state

select(presc_detail, doc_id, nppes_provider_city) %>% filter(doc_id == 'SR7876440') # No other city
select(presc_detail, doc_id, nppes_provider_city) %>% filter(doc_id == 'ZH4066278') # No other city
select(presc_detail, doc_id, nppes_provider_city) %>% filter(doc_id == 'PI6937446') # No other city
select(presc_detail, doc_id, nppes_provider_city) %>% filter(doc_id == 'FP6297812') # No other city

select(presc_detail, doc_id, nppes_provider_city, nppes_provider_state) %>%
  filter(nppes_provider_city == "JAIPUR") # No other state
select(presc_detail, doc_id, nppes_provider_city, nppes_provider_state) %>%
  filter(nppes_provider_city == "TORONTO") # No other state
select(presc_detail, doc_id, nppes_provider_city, nppes_provider_state) %>%
  filter(nppes_provider_city == "ST. JOHN'S") # No other state
select(presc_detail, doc_id, nppes_provider_city, nppes_provider_state) %>%
  filter(nppes_provider_city == "DARTMOUTH") # No other state
